refactor tab-indentation to use formatIndent and formatOutdent event-types
before
- TabIndentation was a class with static methods
- user passed keyboard-event to TabIndentation.handleEvent(event)

after
- TabIndentation is a compound-function
- the main-function accepts a html-element and adds input- and keydown-bindings
- this way you can initialize the bindings with
<tm-textarea ref={TabIndentation}/> which is pretty neat
- TabIndentation.onKeyDown triggers a formatIndent/formatOutdent event
- TabIndentation.onInput checks for input-events of type formatIndent/formatOutdent and then handles these accordingly

this way third-party could be aware of the tab insertions (could be handy for collaborative editors)

thanks @chee for the event-type tip!

NOTE
I think this could be a nice api to follow for other bindings: have a main-function to bind directly with the element and provide its internal functions so that other people can frankenstein tools on top of it.