fix 'Illegal reassignment of import "TabIndentation"'
apparently <tm-textarea ref={TabIndentation}/> tried to assign the ref to TabIndentation instead of calling it, but it only appeared in build. seems a solid-transform bug, wasn't able to easily create a reproduction in the solid playground.

renamed TabIndentation to tabIndentation and the examples to <tm-textarea ref={element => tabIndentation(element)}/>.

added the bug as an issue https://github.com/solidjs/solid/issues/2313 will revert this commit once this is fixed